zooni-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
twilio==9.2.3
cryptography==43.0.1
